+++ src/main/java/cc/cassian/raspberry/config/ModConfig.java
@@ -65,6 +65,7 @@ public class ModConfig {
     public boolean showMusicToast = false;
     public boolean disableFaucetSourceBlocks = false;
     public boolean betterJukeboxes = true;
+    public double jukeboxDistance = 64.0;
 
     public static void load() {
         if (!Files.exists(configPath())) {